# perf: tform: remove default-enabled compare count