For scripts which are dominated by level-0 sorting, counting the compares (and not even printing the result) leads to a large performance impact.

Disabled the counting by default.

The per-thread counter still leads to a performance impact, though it is smaller.

---------

This counter was introduced in f1b83ae, which adds the partial binary search to SplitMerge. There a 1.5% performance improvement is claimed for mincer. Possibly this was determined with a single-threaded test?

This change (default: no counter) compared to current master:

| Test | master (s) | Speedup |
|------|--------------|---------|
|trace|5.09|4.51x|
|mincer|70.57|1.28x|
|mass-fact|59.38|1.12x
|forcer|122.23|1.02x|
|mbox1l|2.73|1.00x|

`forcer` and `mbox1l` tests spend a lot of time sorting polyratfun, so they have much less level-0 sorting time.

Please try your own performance tests and let me know if this is observed on other systems. Here I use 24 threads.