@magv 2025-08-19
That's a crazy improvement for such a small change. Well done finding it.

Why is `form` 10% slower though?