@jodavies 2025-08-19
Ah, looks like fu.frm is basically what I've called "trace" (it comes from aap1000 also). Your slowdown for form is unexpected, though? I don't observe this.