@jodavies 2025-08-19
I modified the per-thread counters to 1024. I don't want to add to AT with an ifdef because the nested padding macros are a pain, and this is not something that should normally be enabled.