@jodavies 2025-08-19
If I re-run the forcer test with `polyratfunexpand=10` then there is a 10% improvement there also. I think we can merge this in the current state.

We should think about a standard set of performance regression tests which are not too time consuming, probably.